check others dead added

+++ includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: andrferr <andrferr@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/28 12:56:41 by andrferr          #+#    #+#             */
-/*   Updated: 2023/01/12 11:06:28 by andrferr         ###   ########.fr       */
+/*   Updated: 2023/01/12 11:57:55 by andrferr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,22 +42,22 @@ typedef struct s_info
 	t_philo	*philos;
 }	t_info;
 
-t_info		*get_info(char **argv);
-int			atoi(const char *str);
-int			check_argc_validity(char **argv, int argc);
-int			check_isdigit(int num);
-int			nbr_of_digits(long n);
-int			philo_init(t_info *info);
-void		*philo_life(void *p);
-void		go_sleep(int t);
-long long	timestamp(void);
-void		take_fork_msg(t_philo *philo);
-void		eating_msg(t_philo *philo);
-void		sleeping_msg(t_philo *philo);
-void		thinking_msg(t_philo *philo);
-void		dead_msg(t_philo *philo);
-void		eating(t_philo *philo);
-void		sleeping(t_philo *philo);
-void		clean_info(t_info *info);
+t_info			*get_info(char **argv);
+int				atoi(const char *str);
+int				check_argc_validity(char **argv, int argc);
+int				check_isdigit(int num);
+int				nbr_of_digits(long n);
+int				philo_init(t_info *info);
+void			*philo_life(void *p);
+void			go_sleep(int t);
+unsigned long	timestamp(void);
+void			take_fork_msg(t_philo *philo);
+void			eating_msg(t_philo *philo);
+void			sleeping_msg(t_philo *philo);
+void			thinking_msg(t_philo *philo);
+void			dead_msg(t_philo *philo);
+void			eating(t_philo *philo);
+void			sleeping(t_philo *philo);
+void			clean_info(t_info *info);
 
 #endif
